requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-ratelimit==4.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.10
numpy==2.3.3
packaging==25.0
psycopg2-binary==2.9.10
pyarrow==21.0.0
pylance==0.36.0
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
django-cors-headers == 4.3.1
djangorestframework-simplejwt == 5.4.0